feat: Use debian.copyright.Copyright module to parse Copyright files to get package licences
Note that not all copyright files are machine readable. If the copyright file is not machine readable,
then we will skip the file and move on to the next package if --ignore-copyright-parsing-errors is set.
If --ignore-copyright-parsing-errors is not set, then we will raise an exception and exit.
If the copyright file is not found, then we will skip the file and move on to the next package if
--ignore-copyright-file-not-found-errors is set. If --ignore-copyright-file-not-found-errors is not set,
then we will raise an exception and exit.

`cpc-sbom --ignore-copyright-parsing-errors --ignore-copyright-file-not-found-errors` will ignore missing
copyright files and ignore copyright parsing errors and will proceed to produce an SBOM.