refactor: Refactor the installed packages file licenses and refactor file lists
Catching UnboundError was not a nice approach. Instead get the installed
package file checksums and add the license if it exists.

Use uuid as a unique identifier for each file in filelist

Ensure that there is an apt cache before creating  SBOM

Generate SBOM as suc if files list is included to permit reading all files on disk to create checksums for the SBOM

Do not use md5sums files for checksums. Instead generate the checksums when generating the SBOM data