fix: Support older versions of python-debian with different Copyright class init methods
For Bionic and older the versions of python-debian is old enough that the Copyright class
init method does not include the strict kwarg.

As such trying to build SBOMs on bionic or Xenial will fail. We can remove this kwarg as
we are catching the NotMachineReadableError exception anyway.

I have tested this successfully in Xenial, Bionic and Focal VMs