feat: add command to inject artifact name
Some fields need to contain the artifact name, but at the point where we
generate the SBOM for an image, we still don't know what the final image
filename will be. When generating an SBOM, the document name is used in
these fields instead.

This adds functionality so the full artifact name can be injected
in these fields later on.